refactor tn handler (#16135)
1. get rid of unnecessary data conversion, like `CreateDatabase -> CreateDatabaseReq`
2. remove redundant codes, like duplicated `NewReplayBaseEntry`
3. use reflect package to make request handling more concise
4. split `rpc/handle.go` into several parts. It's more organized and not a daunting gigantic monster any more

Approved by: @zhangxu19830126, @XuPeng-SH
