+++ pkg/catalog/types.go
@@ -15,6 +15,7 @@
 package catalog
 
 import (
+	"fmt"
 	"strings"
 
 	"github.com/matrixorigin/matrixone/pkg/container/types"
@@ -402,9 +403,9 @@ type CreateDatabase struct {
 	Name        string
 	CreateSql   string
 	DatTyp      string
-	Owner       uint32
-	Creator     uint32
-	AccountId   uint32
+	Owner       uint32 // roleid
+	Creator     uint32 // userid
+	AccountId   uint32 // tenantid
 	CreatedTime types.Timestamp
 }
 
@@ -431,12 +432,9 @@ type CreateTable struct {
 	Defs         []engine.TableDef
 }
 
-type UpdateConstraint struct {
-	DatabaseId   uint64
-	TableId      uint64
-	TableName    string
-	DatabaseName string
-	Constraint   []byte
+func (t CreateTable) String() string {
+	return fmt.Sprintf("{aid-%v,uid-%v,rid-%v}: %d-%s:%d-%s, %q",
+		t.AccountId, t.Creator, t.Owner, t.DatabaseId, t.DatabaseName, t.TableId, t.Name, t.CreateSql)
 }
 
 type DropOrTruncateTable struct {
