+++ pkg/pb/api/api.go
@@ -144,6 +144,14 @@ func (m *SyncLogTailResp) UnmarshalBinary(data []byte) error {
 	return m.Unmarshal(data)
 }
 
+func (m *TNStringResponse) MarshalBinary() ([]byte, error) {
+	return m.Marshal()
+}
+
+func (m *TNStringResponse) UnmarshalBinary(data []byte) error {
+	return m.Unmarshal(data)
+}
+
 func (m *PrecommitWriteCmd) MarshalBinary() ([]byte, error) {
 	return m.Marshal()
 }
