+++ pkg/vm/engine/tae/catalog/table.go
@@ -113,27 +113,21 @@ func NewTableEntryWithTableId(db *DBEntry, schema *Schema, txnCtx txnif.AsyncTxn
 }
 
 func NewSystemTableEntry(db *DBEntry, id uint64, schema *Schema) *TableEntry {
-	opts := btree.Options{
-		Degree: 4,
-	}
-	e := &TableEntry{
-		ID: id,
-		BaseEntryImpl: NewBaseEntry(
-			func() *TableMVCCNode { return &TableMVCCNode{} }),
-		db:         db,
-		TableNode:  &TableNode{},
-		link:       common.NewGenericSortedDList((*ObjectEntry).Less),
-		entries:    make(map[types.Objectid]*common.GenericDLNode[*ObjectEntry]),
-		deleteList: btree.NewBTreeGOptions(DeleteEntry.Less, opts),
-		Stats:      common.NewTableCompactStat(),
-	}
+	e := NewReplayTableEntry()
+	e.ID = id
+	e.db = db
+
 	e.TableNode.schema.Store(schema)
 	e.CreateWithTS(types.SystemDBTS, &TableMVCCNode{Schema: schema})
+
 	var sid types.Uuid
-	if schema.Name == SystemTableSchema.Name {
-		sid = SystemObject_Table_ID
-	} else if schema.Name == SystemDBSchema.Name {
+	if schema.Name == SystemDBSchema.Name {
+		if DefaultTableDataFactory != nil {
+			e.tableData = DefaultTableDataFactory(e) // TODO(aptend): add data handle
+		}
 		sid = SystemObject_DB_ID
+	} else if schema.Name == SystemTableSchema.Name {
+		sid = SystemObject_Table_ID
 	} else if schema.Name == SystemColumnSchema.Name {
 		sid = SystemObject_Columns_ID
 	} else {
@@ -149,12 +143,12 @@ func NewReplayTableEntry() *TableEntry {
 		Degree: 4,
 	}
 	e := &TableEntry{
-		BaseEntryImpl: NewReplayBaseEntry(
-			func() *TableMVCCNode { return &TableMVCCNode{} }),
-		link:       common.NewGenericSortedDList((*ObjectEntry).Less),
-		entries:    make(map[types.Objectid]*common.GenericDLNode[*ObjectEntry]),
-		deleteList: btree.NewBTreeGOptions(DeleteEntry.Less, opts),
-		Stats:      common.NewTableCompactStat(),
+		BaseEntryImpl: NewBaseEntry(func() *TableMVCCNode { return &TableMVCCNode{} }),
+		link:          common.NewGenericSortedDList((*ObjectEntry).Less),
+		TableNode:     &TableNode{},
+		entries:       make(map[types.Objectid]*common.GenericDLNode[*ObjectEntry]),
+		deleteList:    btree.NewBTreeGOptions(DeleteEntry.Less, opts),
+		Stats:         common.NewTableCompactStat(),
 	}
 	return e
 }
