+++ pkg/vm/engine/tae/iface/txnif/types.go
@@ -282,7 +282,7 @@ type TxnStore interface {
 	GetRelationByID(dbId uint64, tid uint64) (handle.Relation, error)
 
 	CreateDatabase(name, createSql, datTyp string) (handle.Database, error)
-	CreateDatabaseWithID(name, createSql, datTyp string, id uint64) (handle.Database, error)
+	CreateDatabaseWithID(ctx context.Context, name, createSql, datTyp string, id uint64) (handle.Database, error)
 	GetDatabase(name string) (handle.Database, error)
 	GetDatabaseByID(id uint64) (handle.Database, error)
 	DropDatabase(name string) (handle.Database, error)
